test(config): stop the compat-entry logging test racing the tracing interest cache
